# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "HeyPizza",
    "description": "...",
    "version": "1.0.0"
  },
  "host": "localhost:3101",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/slack/verification": {
      "post": {
        "tags": [
          "Slack"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "challenge": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/slack/webhooks": {
      "post": {
        "tags": [
          "Slack"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "any"
                },
                "challenge": {
                  "example": "any"
                },
                "event": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/slack/new-feedback": {
      "post": {
        "tags": [
          "Slack"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "example": "any"
                },
                "user_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}